requirements.txt
aiosmtpd==1.4.6
alembic==1.13.3
atpublic==5.0
attrs==24.2.0
babel==2.16.0
blinker==1.8.2
cachetools==4.2.4
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-api-core==1.34.1
google-auth==1.26.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.48.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
langdetect==1.0.9
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
WTForms==3.1.2
